+++ methods_00_trace.go
@@ -13,9 +13,7 @@ func (l *Logger) Trace(args ...any) {
 
 	l.logWithLabel(
 		l.labelTrace(),
-		uint32(
-			helpers.GetEstimatedMessageSize("", args),
-		),
+		helpers.GetEstimatedMessageSize("", args),
 		args,
 	)
 }
@@ -28,9 +26,7 @@ func (l *Logger) Tracef(format string, args ...any) {
 	l.logfWithLabel(
 		l.labelTrace(),
 		format,
-		uint32(
-			helpers.GetEstimatedMessageSize(format, args),
-		),
+		helpers.GetEstimatedMessageSize(format, args),
 		args,
 	)
 }
@@ -47,9 +43,7 @@ func (l *Logger) Tracew(msg string, keysAndValues ...any) {
 	l.logwWithLabel(
 		l.labelTrace(),
 		msg,
-		uint32(
-			len(msg)+helpers.GetEstimatedMessageSize("", keysAndValues),
-		),
+		uint32(len(msg))+helpers.GetEstimatedMessageSize("", keysAndValues),
 		keysAndValues,
 	)
 }
